+++ dbdaora/conftest.py
@@ -11,3 +11,22 @@ async def dict_repository(mocker, dict_repository_cls):
         fallback_data_source=DictFallbackDataSource(),
         expire_time=1,
     )
+
+
+@pytest.fixture
+def async_iterator():
+    return AsyncIterator
+
+
+class AsyncIterator:
+    def __init__(self, seq):
+        self.iter = iter(seq)
+
+    def __aiter__(self):
+        return self
+
+    async def __anext__(self):
+        try:
+            return next(self.iter)
+        except StopIteration:
+            raise StopAsyncIteration