+++ dbdaora/hash/_tests/test_unit_hash_service.py
@@ -10,6 +10,7 @@ def service():
     s = HashService(
         repository=asynctest.MagicMock(id_name='id'),
         circuit_breaker=asynctest.MagicMock(),
+        fallback_circuit_breaker=asynctest.MagicMock(),
         cache={},
     )
     s.entity_circuit = asynctest.CoroutineMock()
@@ -41,12 +42,16 @@ async def test_should_get_already_not_found_when_getting_one(service):
 
 
 @pytest.mark.asyncio
-async def test_should_set_cache_entities_not_found_when_getting_many(service):
+async def test_should_set_cache_entities_not_found_when_getting_many(
+    service, async_iterator
+):
     fake_entity = {'id': 'fake'}
-    service.entities_circuit.return_value = [fake_entity]
+    service.repository.query.return_value.entities = async_iterator(
+        [fake_entity]
+    )
     service.repository.id_name = 'id'
 
-    entities = await service.get_many('fake', 'fake2', 'fake3')
+    entities = [e async for e in service.get_many('fake', 'fake2', 'fake3')]
 
     assert entities == [fake_entity]
     assert service.cache['fake2'] == CACHE_ALREADY_NOT_FOUND
@@ -60,9 +65,9 @@ async def test_should_get_already_not_found_when_getting_many(service):
     service.cache['fake2'] = CACHE_ALREADY_NOT_FOUND
     service.cache['fake3'] = CACHE_ALREADY_NOT_FOUND
 
-    entities = await service.get_many('fake', 'fake2', 'fake3')
+    entities = [e async for e in service.get_many('fake', 'fake2', 'fake3')]
 
-    assert not service.entities_circuit.called
+    assert not service.repository.query.called
     assert entities == [fake_entity]
     assert service.cache['fake2'] == CACHE_ALREADY_NOT_FOUND
     assert service.cache['fake3'] == CACHE_ALREADY_NOT_FOUND