+++ dbdaora/hash/repositories/_tests/test_integration_repository_aioredis_hash_get_many.py
@@ -2,18 +2,12 @@ import asynctest
 import pytest
 from jsondaora import dataclasses
 
-from dbdaora import HashQueryMany
-from dbdaora.exceptions import EntityNotFoundError
-
 
 @pytest.mark.asyncio
 async def test_should_set_already_not_found_error_when_get_many(
     repository, mocker
 ):
     fake_entity = 'fake'
-    expected_query = HashQueryMany(
-        repository, fake_entity, memory=True, many=[fake_entity]
-    )
     repository.memory_data_source.hgetall = asynctest.CoroutineMock(
         return_value=None
     )
@@ -26,10 +20,10 @@ async def test_should_set_already_not_found_error_when_get_many(
     repository.memory_data_source.set = asynctest.CoroutineMock()
     repository.memory_data_source.hmset = asynctest.CoroutineMock()
 
-    with pytest.raises(EntityNotFoundError) as exc_info:
-        await repository.query(many=[fake_entity]).entities
+    assert [
+        e async for e in repository.query(many=[fake_entity]).entities
+    ] == []
 
-    assert exc_info.value.args == (expected_query,)
     assert repository.memory_data_source.hgetall.call_args_list == [
         mocker.call('fake:fake'),
     ]
@@ -53,7 +47,9 @@ async def test_should_get_many_from_fallback(repository, fake_entity):
         fake_entity
     )
 
-    entities = await repository.query(many=[fake_entity.id]).entities
+    entities = [
+        e async for e in repository.query(many=[fake_entity.id]).entities
+    ]
 
     assert entities == [fake_entity]
     assert repository.memory_data_source.exists('fake:fake')
@@ -71,7 +67,12 @@ async def test_should_get_many_with_one_item_already_not_found_from_fallback(
         fake_entity
     )
 
-    entities = await repository.query(many=[fake_entity.id, 'fake2']).entities
+    entities = [
+        e
+        async for e in repository.query(
+            many=[fake_entity.id, 'fake2']
+        ).entities
+    ]
 
     assert entities == [fake_entity]
     assert repository.memory_data_source.exists('fake:fake')
@@ -91,9 +92,12 @@ async def test_should_get_many_with_one_item_already_not_found_and_another_not_f
         fake_entity
     )
 
-    entities = await repository.query(
-        many=[fake_entity.id, 'fake2', 'fake3']
-    ).entities
+    entities = [
+        e
+        async for e in repository.query(
+            many=[fake_entity.id, 'fake2', 'fake3']
+        ).entities
+    ]
 
     assert entities == [fake_entity]
     assert repository.memory_data_source.exists('fake:fake')