+++ dbdaora/sorted_set/service/__init__.py
@@ -1,4 +1,4 @@
-from typing import Any, Optional, Sequence
+from typing import Any, AsyncGenerator, Optional, Sequence
 
 from cachetools import Cache
 
@@ -8,16 +8,18 @@ from ..entity import SortedSetData, SortedSetEntity
 
 
 class SortedSetService(Service[SortedSetEntity, SortedSetData, FallbackKey]):
-    async def get_many(self, *ids: str, **filters: Any) -> Sequence[Any]:
+    def get_many(
+        self, *ids: str, **filters: Any
+    ) -> AsyncGenerator[SortedSetEntity, None]:
         raise NotImplementedError()  # pragma: no cover
 
-    async def get_many_cached(
+    def get_many_cached(
         self,
         ids: Sequence[str],
         cache: Cache,
         memory: bool = True,
         **filters: Any,
-    ) -> Sequence[Any]:
+    ) -> AsyncGenerator[SortedSetEntity, None]:
         raise NotImplementedError()  # pragma: no cover
 
     async def delete(